Move state for the currently edited game and add-apps up a level (from RightBrowseSidebar to BrowsePage). Fixed bug where saving add-apps while there was exactly one for tha platform already would delete the already present one.
How games are being save has changed:
Previously you could only edit games and not add new ones, so the logic that saves games only found where that game originated from (based in its ID) and overwrote that game and then saved the platform it belongs to. If the game wasnt found it would throw an error and crash the application.
To make saving new games possible, the logic now does the same thing except for when the game is not found it instead try to save it to a platform with a name that matches the games "platform" property. If that is not found either, it will save it to a platform called "Unknown Platform" (which will be automatically created if it does not exist).
Note: There is no interface for creating new games yet, but this should bring it one step closer to working! :D