Bug fixes (mostly) related to creating new games.
Change type of IGameInfo#dateAdded from number to string.
(this is because converting a date to number (time stamp) discards its timezone, which its needs to know when it saves it later on)

Copy the "dateAdded" property from parsed games to raw games when saving.
(this was previouslly disabled since you couldnt change this property, and since it will shave off the 3 last digits of the milliseconds)

When saving changes made to a game, update its "orderTitle" (this is to prevent the games from being ordered by their previous title, or no title in the case of new games).

After a new game is created, it is now automatically selected in the browser list/grid.