feat: add a "clear search" button to the search bar
Show an icon in the left side of the search bar. If the search bar has
no text, the icon will be a magnifying glass. If there is text in it,
the icon will be a circle with an x inside, and clicking it will clear
the search bar (and "reset" the current search).