build: make the "build" and "watch" gulp tasks faster and simpler
The biggest change is that the gulp tasks use "child_process.exec"
instead of the gulp-ified wrapper modules. This makes them way faster.
It does however make the console output less pleasant.
The "renderer" folder in the "app.asar" file is now renamed to "window".
This is because "tsc" outputs files in the "renderer" folder now.