feat: each playlist is now only shown in its respective "game library"
Playlists now have a new optional field "library". It is a string. If
it is set, the playlist will be shown only when the "game library" with
a matching "route" is selected. If the field is not set, it will default
to being shown in the "default" game library.
New playlists will have their "library" field set to match the current
"game library".