feat: search is now limited to the selected "game library" (& more)
Searching will only display games of the selected, or most recently
selected, "game library". It will default to the "default" game library.
The route of the last selected game library is also stored in the
preferences file. This makes the "most recently selected" feature above
carry over between sessions.
When creating new games the prefix of the selected game library will be
taken into account when finding/creating the platform XML to save the
game to.