feat: add a "preview box" where you can view, add and remove game images
While editing, the screenshot in the right sidebar will be replaced with
a "preview box" where both the thumbnail and screenshot can be seen.
If the thumbnail/screenshot is missing, a text will say so and a button
will appear that lets you browse to and add an image.
You can also drag-and-drop a file onto the "preview" to add it.
If the image is found, a preview of it will be shown and a red "trash"
icon will appear that deletes the image if clicked twice.
Also fixed a minor bug: the "&" character was not getting replaced with
underscores when parsing game titles to image filenames.