feat: make certain add-app application paths act as special features
Additional Applications that has the application path ":message:" or
":extras:" will NOT be launched like the others.
The add-apps with ":message:" will have a popup appear with the value of
the add-apps Command Line property as its body.
The add-apps with ":extras:" will open a folder relative to the path
"<FlaspointPath>/Extras/". The folders relative path is equal to the
add-apps Command Line property.

Both of these used to be batch files that came with Flashpoint, but now
they are built in. ":" is not a legal filename character, so this these
will now "shadow" any actual files.