feat: use "filename" instead of "platform" when selecting image folder
The games has an additional property ("filename") which is set to the
filename of the last file the game was loaded from or saved to. This
property is used to determine which image folder will be used for the
game, instead of the "platform" property.

The reason for this change is that the filename is a combination of both
the platform property and the library prefix.