fix: make adding game images more stable
Adding images with the "Add Screenshot" and "Add Thumbnail" buttons now
create the image folders if they are missing and load them to the cache.
The buttons are now disabled if the current image folder name is empty
(this occurs for new games without a platform set).
Saving game images will now use the games ID in the filename (instead of
using the title).
Attempting to cache missing image folders now logs a warning in the
console instead of letting it throw and error.