tweak: load game images using their "id" before their "title"
Previously the launcher would first try to load a games image using it's
title, and if that failed it tried with the id, and if that failed it
gave up (title -> id -> give up).
Now it does that same thing, but it tries the title and id in the
reverse order (id -> title -> give up).
This should increase performance as long as you have more images named
after the games id than those named after the title.