build: update tsconfig structure
The location and contents of the "tsconfig.json" has been changed.
Instead of having all tsconfig files in the root folder of the project,
there is only one file at the root and the rest is spread out in the
"src" folder.
This is because the typescript compiler will by default look for which
tsconfig file to use by looking in the current folder and then going up
one folder recursively until it is found.
The benefit of this setup is that VS Code can now find the tsconfig
files too (instead of just using one across the whole project).