fix: fix naming error (and related error) in log inter-process api
Rename the parameter "length" to "start" (of the "refresh" function).
Since that is what the argument value will be used as.
Also make the renderer always request all log entries when created
(I don't know why I made it use the length of the local entries array,
since it will crash down the line if it is anything other than 0).