feat: add a "debounce" utility function
This is a utility function that you can use to wrap event listeners with
if they receive identical events multiple times in short periods of
times.
An example of when this is useful is for the "change" event of node's
file watcher. Each time a file is changed, this event is emitted two or
three times with identical arguments a few milliseconds apart.
By using this function you can "buffer" the first event, reject all
identical events sent within a few milliseconds, then release the first
event once.