feat: improve FolderWatcher (recursive watch, file stats & more)
FolderWatcher has been improved in many ways.
It can watch folders recursively (with variable depth).
It automatically fetches and stores the "stats" of the watched files.
It "debounces" the change events, meaning that it will now only emit one
event per change (instead of 2-3).
It uses an event queue to queue edits made to its file related vars
(this removes some very unlikely race conditions).
It applies a "hot fix" to the underlying node file watcher.
It has types for the "once" method in addition to "on".
And some more minor stuff.