feat: improve ThemeManager (and related code in other classes & files)
ThemeManager has been improves so it should correctly add and remove
themes as well as correctly emit "change" events.
Selecting a theme in the drop-down (in ConfigPage) will now focus the
input field of the drop-down element.
FolderWatcher no longer emits "rename" events (this could be brought
back later, but under it's own event instead of sharing it with
"change").