feat: add "Open File Location" context option to game elements
Game elements (GameGridItem & GameListItem) now have a context menu with
one item ("Open File Location").
When the item is clicked, the launcher will try to locate the file the
game's "launch command" is pointing at, and then open the folder
containing the file (in the OS's file explorer) and highlight/select
the file. If the launcher fails to "extract" the path of the game, or if
the file does not exist, an error dialog will appear and inform the user
of what went wrong.
Also add a developer tool for testing the