refactor: continue reorganizing css variables (& add lost css)
All "--theme" CSS variables has been renamed/reorganized to "--layout"
variables.
Some CSS that got lost when splitting the original stylesheet into two
files has been recovered and inserted into the right place (it's related
to the scrollbar).
Both CSS files now have a comment at the top of the file that gives a
short explanation of the purpose of the file.