refactor: change and apply new "conventions" for components
Almost all components have been changed to various degrees to make them
(among other things) easier to read and understand.
These changes include (but are not limited to):
- Add comments to almost all prop and state properties
- Remove the "private" keyword from all(?) properties of components
- Convert many interfaces into types
- Rewrite some class components as function components (with hooks)
- Remove the character "I" from the beginning of prop and state names
- Remove unnecessary constructors from components
- Change to a new "convention" for indenting JSX/TSX (read more below)

The new "convention" for JSX/TSX is very similar to the previous one,
the only thing that changed is how to indent props.
The new "convention" can be explained like this:
All React Elements (<a>these things</a>) with more than one prop
should separate all props with a new line. The props on these lines
should be indented one "step" further than the React Element tag(s).
Example of React Element before this "convention" (not a real example):
<div className='link-container' ref={containerRef}
     onClick={onContainerClick}>
  <a href='/home'>Home</a>
  <a href='/browse'>Browse</a>
</div>
The same React Element but with the new "convention":
<div
  className='link-container'
  ref={containerRef}
  onClick={onContainerClick}>
  <a href='/home'>Home</a>
  <a href='/browse'>Browse</a>
</div>