feat: add curate page, implement loading and displaying of a curation
A new page/tab called "Curate" has been added.
On the page you can load any number of curations, which are then
displayed on that page (all meta fields, a list of all files and folders
inside the content folder, the thumbnail and the screenshot).
The displayed meta fields can be edited.