feat: add "context reducer" and use it to store the curate page's state
A new component has been added (ContextReducerProvider).
It stores the state of a reducer (using the "useReducer" hook) and
provides it and the dispatch function through a context.
This makes it easy to decouple the state from another component and move
it up the element tree.
The curate page's state (of the currently loaded curations) has been
moved to a context. The provider is located close to the app's root,
making the state survive page changes.