feat: improve the curate page (remove button, index content & more)
Add a "Remove" button to each CurateBox that removes the curation when
clicked.
Add a list of all the content files/folders (and if they will collide
with any existing file/folder) to each CurateBox. The list also contains
the size of each file that will be imported (and the files they
"collide" with).
If any content files/folders of a curation will "collide", a count of
how many will collide (compared to the total number of content
files/folders) is shown above the list.