feat: make curation importer import the game info and images
Pressing the "import" button in a CurationBox now imports the curation's
meta data and images (by creating and inserting a game entry and
copying/extracting the thumbnail and screenshot images).
The code that saves games and images has been refactored to make it
usable by both the curation importer and the right sidebar (editor).