feat: make the "import" button extract the content files
Pressing the "import" button in a CurationBox now extracts the files and
folders inside the "content" folder into the "htdocs" folder. It also
removes the curation afterwards, since it is done importing.
Add a general purpose unzip function (which is used to unzip the
content).
Also improve the performance of the curation reducer.