feat: improve the curation importer in various ways
Create a list of warnings for each curation (at the moment there is only
a warning for application paths that don't use the "http" protocol).
Display a list of all warnings at the bottom of each Curation Box.
Make the "Extreme" field use a checkbox instead of an input field.
Make the "Genre", "Platform" and "Application Path" fields use drop-down
input fields (with the suggestions pulled from the existing games).
Show a "zoom in" cursor while hovering over a thumbnail or screenshot.
+ some minor refactoring