feat: improve ObjectParser (add ".value", add docs, rewrite some code)
ObjectParser, and the associated classes/types/functions, has gotten
new and improved documentation (not all documentation was rewritten
though).
Some of the code has been rewritten or reformatted to look better
(although barely).
A new ".value" property has been added to ObjectParserProp, this makes
it possible to directly access the wrapped value (this will be used in
the upcoming usage of this class for "parsing" curation meta).