feat: support curating add-apps, use ObjectParser for curations & more
Additional Application meta data in curation meta is now extracted,
displayed in the curation importer and saved when imported.
ObjectParser is used to convert the "raw" curation meta data into a more
programmer friendly format (instead of the previous basic function).
Divider elements have been added to the CurateBox component to make it
easier to see where each "section" begins and ends.
You can no longer preview an CurateBoxImage that does not have an image
(the cursor does not change to "zoom-in" when hovering over one of
these).