feat: make EventQueue#push optionally return a promise (& more)
A second parameter has been added to EventQueue#push which decides if it
should return a promise. This promise is resolved right after the event
function is executed (or resolved), or rejected if the event function
throws an error (or rejects).
The EventFunction type has been expanded to include async functions.
These have always been supported, the type did just not include them.