feat: lock and disable all input in CurationBox while importing (& more)
The curations can now be "locked". While locked, all input elements used
to edit its meta are disabled.
GameManagerPlatform now uses an EventQueue to queue attempts to save the
data to file.
The data is converted to XML immediately, and only the writing to the
file is put on the queue. This should make saving more robust and less
prone to silently failing (and possibly losing any changes made).
InputField and DropdownInputField now use the "disabled text color"
while disabled.
The drop down of DropdownInputField can no longer be expanded or
operated while disabled.