feat: add new warnings, and improve the way warnings are displayed
A warning is now shown when "Release Date" is incorrectly formatted.
Warnings are now shown when "Genre", "Platform" or "Application Path"
has values that are not already suggested.
The files list in CurateBox now has a max height, and shows scrollbars
if the content doesn't fit inside it.
The warning messages inside CurateBoxWarnings no longer show the name
of the warning (it was a waste of space).