fix: update parser to work with new curation meta format (& bug fix)
The new curation format has been changed to not use any "lists" and
instead separate the values by semicolons, the parser has been changed
to reflect this change.
The parser incorrectly tried to grab the value from "Language" instead
of "Languages" (the "s" was missing).
The order in which the values are parsed have been changed to
alphabetical order.