feat: add default values to "Platform" and "Application Path"
When loading a curation, if the "Platform" and/or "Application Path"
fields were missing from the meta, they are automatically set to default
values.
If the "Platform" field is missing, it's set to "Flash".
If the "Application Path" field is missing, it's set to the most used
application path for games with the same platform.