refactor: rewrite much of the code related to config and preferences
A lot of the code related to both config and preferences have been
refactored.
The config and preferences APIs have been changed to be more similar.
This includes changes such as:
* The config "back end" is now its own instance of a class, instead of
  being spread all over the place.
* Making the "back end" of the config API solely responsible for the
  config file (the renderer now has to send a request to save to it over
  IPC, instead of being able to just save to it directly). This is how
  the preferences have always done it.
* Method names, errors and comments have been changed to be more similar
  across the config and preferences code.
A lot of other changes have been made too, but it would take forever to
list them.

Also, the config "startRouter" has been renamed to "startServer".