feat: add ObjectParserPropNone and 3rd argument to ObjectParserProp#map
If ObjectParserProp's method "prop" is called on a property that can not
be found, it now returns an instance of the new class
ObjectParserPropNone (instead of an instance of ObjectParserProp). This
class is very similar to ObjectParserProp, but with the difference that
it doesn't wrap any object and thus it doesn't run any code (other than
returning values of the expected types). This way it uses less
performance, and doesn't log redundant errors.
The property "value" of ObjectParserProp has been removed in favor of
adding a 3rd parameter to its methods "map" and "mapRaw" which is the
map that is being iterated over. This enables the code that uses
ObjectParser to still access the raw value, and makes the methods more
similar to the "array" and "arrayRaw" methods.