feat: add build version system (each build gets a version number)
A file named ".version" (containing a time-stamp of when the build was
started) will be created for each build and release build.
When building (in general), the file will be created in the projects
root folder (next to "package.json").
When packing release builds, the file will be created in the same folder
as the launchers executable.
The launcher will try to read the value of this file on startup, format
it to a YYYY-MM-DD date and display it as the "current version" instead
of "alpha" (both in the windows title-bar and the about page).