perf: search for a subset of valid image filenames first (& more)
Speed up the mapping of most valid image filenames by first searching
for a small subset of valid filenames before doing a full search.
Replace the use of "path.join" with a function that just adds the
folder and filenames strings together, with a slash in between.

This is over 100 times faster on Flashpoint 6.2 on my machine.