feat: add "export buttons" to the browser page context menu
Add two new options to the context menu in the browse page.
The first option is "Export (Meta)" which opens a save dialog and lets
you select where to export the games meta (curation format).
The second option is "Export (Meta & Images)" which opens a open dialog
and lets you select what folder to save the meta and images in.
Both options check if any file has to be overwritten, and asks for
permission before overwriting.