refactor: generalize DropDown so it can be used outside of the log page
Rename the CSS (BEM) block "checkbox-dropdown" to "simple-dropdown".
Move the CSS code related to the specific contents of the dropdown in
the log page from being part of the "checked-dropdown" block to the
"log-page" block (BEM), this makes the dropdown usable in other
scenarios too (it's only used in the log page at the moment).
Use the "onMouseDown" event instead of "onClick" to toggle expansion.