feat: add string formatter function (and add options to variable string)
Add new util function formatString. It is similar to parseVariableString
but it supports non-string arguments, and returns an array instead of a
string as longs as at least one object should be returned.
Add an  optional options argument to the variable string functions. This
enables you to select which open and close characters it should look
for.