refactor: simplify the language container types
The language container types, and the function that create the language
container objects, have been rewritten to derive their values from a
template object.
This was done to reduce the number of times you had to type the name of
a language string (in different parts of the code base) to add a new
language string.
Now you only have to type it once inside the language template object.