refactor: clean up (remove unused code, style changes etc.)
One notable change is that GameImageSplit no longer has the "type" prop,
which was only used to be passed back in the callback props.
The purpose was to make it possible to have a single callback for
multiple images and use the value of "type" to know which GameImageSplit
called the callback.