refactor: move more stuff to the back process
The configs are now loaded by the back (and sent to main and renderer).
More work on making the games be stored in the back, sent to the
renderer App and down the component tree has been done.

The launcher can now be built and executed. It will however never get
access to all the data necessary for it to work properly (like platforms
and games).