refactor: make browsing and searching games work with the back (& more)
You can now browse and search for games in the BrowsePage (including
ordering).
Search results are cached in the back (the hash of the search object is
paired with the resulting array of games and stored in the state).
Games are fetched in "pages" and cached in the renderer for each
library, and are cleared if the search options for that library changes.