refactor: launch/save/delete games work (& more)
Saving and deleting games now work (it's not well tested so it might not
work as intended under heavy load or if the launcher is shut down while
saving).
Launching games now work (the whole GameLauncher file was moved to the
back, only minor changes has been made to it so it works).
Some work on making playlists usable has also been done.
And some other minor stuff as usual.