refactor: move the log management to the back (& launch add-apps & more)
Move the "log management" from the main to the back process.
Launching add-apps from the back now works (some front-end stuff just
had to allow for it).
Attempt to re-add support for the add-app property "waitForExit" (it was
remove at some point for some reason - this hasn't been tested yet).