refactor: move background services to the back process
Move the background services from the main to the back process.

Note: Stopping the Redirector process can close the websocket. It is
more likely to happen if you run the Redirector before the launcher.
Perhaps it is caused by the websocket being opened through the proxy
and when the proxy setting changes it gets closed?