refactor: move languages and themes to the back
Move language and theme "management" from main to back.
Quitting the Electron application sends a message to the back and waits
for it to respond before quitting.
The back process is now detached from the main process, which gives it
time to exit gracefully (which should take less than a second anyways).
The Redirector/Fiddler is now spawned detached.