feat: rewrite playlists code and move it to the back process
Playlists are now managed by the back process.
A folder watcher loads all playlists and looks for changes to the files
(both made by the launcher itself, or an outside program).
The renderer sends requests to update/add or delete playlists.
Playlist icons are cached into blobs in the renderer, this makes them a
lot more performant.
Playlists now use their filename as their ID, and this can be edited
from the renderer.
The playlist editor in the renderer has been improved to work more like
the game meta data editor.
And a bunch of smaller edits, as usual.