fix: fix various bugs
The back servers now properly tries all ports in their range.
GameManager no longer throws an error if the "Data/Platforms" folder is
missing when loading the platforms.
The application now closes more consistently if something goes wrong
during startup (although it still sometimes does not).
Running out of games to show in RandomGames no longer causes a "crash"
(the back used to include undefined in the array of games when it ran
out of games to choose from).
"parseVariableString" no longer throws an error if "splitVariableString"
returns no split.
Types related to curation that are used outside of the renderer has been
moved to shared.