fix: bugs related to the selected game whenever the search changes
If the search changes and the selected game is no longer in the results,
it is now deselected.
If the search changes and a game is selected, the renderer will now get
the index of that game in the search and request it individually. This
will cause the game grid/list to "scroll" to that game instantly.