feat: add settings for "on demand images" (enable/disable and base url)
Add a new config ("onDemandBaseUrl"). This points to the base url to
download images on demand from. It's here just so the URL is not hard
coded and can be replaced by anyone before a release.
Add a new preference ("onDemandImages"). This determines whether or not
missing images will be downloaded on demand. This can be changed from
the config page.