feat: add meta edit importer/exporter
* Add a component (MetaEditExporter) that is used for selecting which
game properties to include when exporting a meta edit
* Add an option to the context menu for games in the game list/grid
("Export Meta Edit") that opens the meta edit exporter "popup"
* Add a folder ("Flashpoint/Data/MetaEdits") that meta edit files are
exported to and imported from
* Add a developer tool ("Import Meta Edits") that imports all meta edit
files in the meta edits folder

Meta edits are used for making changes to existing games (such as
correcting typos). These used to be written by hand and applied directly
to the XMLs by a separate program. But since the launcher now uses a
database, the tool no longer works and a substitute was needed.