perf: Only request one game for the right sidebar at a time
Problem: The game data in the right sidebar could lag behind the current
selection by a lot if the selection was moving around too quickly for
too long.
Solution: Ensure that BrowsePage only requests one game data at a time
by storing the most recent request (if it is already busy) and
performing it after the current request is done.