fix: Make "dateModified" a valid "order by" & log when it is invalid
Setting the "order by" drop-down value to "Date Modified" would cause it
to silently partially fail. It would set the filter correctly but the
back would fail to set it as the preference and instead set it to the
default ("Title"). This would cause it to seemingly reset to "Title" the
next time the launcher is started.
* Add "dateModified" to the array of valid "GameOrderBy" values.
* Make drop-down use the same array of valid "GameOrderBy" values as
the preferences.
* Make attempts at setting the "order by" or "order reverse" values to
something invalid fail completely and log errors to the console.