fix: Copy read-only content files when importing curation
fs-extra's copy function throws an error and halts copying if it fails
to copy a file. It is also unable to copy read-only files.
Reverting back to the "copyFolder" function, as well as using original
fs functions (over fs-extra functions), for copying files makes it able
to copy read-only files.