feat: add ExecMapping support for non-native Linux platforms.
Add support for applicationpath configuration with execs.json for
non-native Linux platforms. If an override path is not found in
execs.json under the key "wine", keep the old default behavior of
using the unconfigurable WINE launch command.
Previous behavior: if "native" was not checked, the system's WINE
would be used to launch the game. This is not sufficiently configurable,
and the long-term plan will use a portable docker environment, launched
from shell scripts.