fix: manually implement tree kill
This is a first try for making process killing work on mac and linux.
It is so far only tested on linux, but I'll test it on mac and windows
soon.
It turns out that this requires making the kill() and restart()
functions into async's - before, it looks like it was exiting before
the services could get killed, not sure how. Anyway, awaiting some
stuff seems to have fixed that.