Add Wine to PATH if it's bundled in FPSoftware
On Mac & Linux systems, check if the `FPSoftware/Wine/bin` directory exists relative to fpPath. If it does, add it to the PATH, so Wine launches without the user having to modify the PATH themselves.
The PATH will not be modified if the directory doesn't exist, or if the launcher is running on Windows (obviously).