Fix Ruffle not launching on macOS
When Ruffle is downloaded by the launcher on macOS, it is downloaded as an application rather than a standalone binary. This results in the launcher being unable to launch Ruffle, as it expects to find it in `ruffle` instead of its actual location (`Ruffle.app/Contents/MacOS/ruffle`), relative to `Data/Ruffle/standalone/latest`.
This PR corrects that behavior, allowing the launcher to automatically set the proper permissions and launch Ruffle on macOS like it can on Linux.